fix compilation error due to boost::filesystem symbols interacting with wxString
boost::filesystem since version 1.81 throws compilation error:

  call of overloaded ‘_check_convertible_to_path_source(const wxString&)’ is ambiguous

This arises from this piece of code:

  if (m_textDrupalName->GetValue() != wxEmptyString)

I don't know how exactly this happens, but apparently trying to compare two different strings
leads to an attempt to instantiate boost::filesystem::path, and that fails. The solution to
this problem is to avoid importing all boost::filesystem symbols and only explicitly use them
in 4 places where that is necessary.